requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
